Fix incorrect behavior of RandomSurface
The random surface was creating surfaces with the correct distance to the closest neighbor. However, it was not taking the whole surface into account but only local closest neighbors. This means when actually correcting the factor 2 mistake made in the probability calculation the surfaces would no longer be built correctly. The system was self healing before because it could redistribute the coordinates until the **local** densities were correct, but the global density was always incorrect.
With the new changes this problem is now fixed. The interpretation of closest neighbor was extended to closest neighbors. Closest neighbors are now defined as all coordinates within 1.5 times the desired average distance.